+++ src/test/java/com/github/berkbavas/breakout/math/Matrix2x2Test.java
@@ -6,32 +6,32 @@ import org.junit.Test;
 import java.util.Optional;
 
 public class Matrix2x2Test {
-    private static final float EPSILON = 1e-8f;
+    private static final double EPSILON = 1e-8f;
 
     @Test
     public void testDeterminant() {
-        final float m00 = -1.13f;
-        final float m01 = -3.58f;
-        final float m10 = 3.5f;
-        final float m11 = 21.24f;
+        final double m00 = -1.13f;
+        final double m01 = -3.58f;
+        final double m10 = 3.5f;
+        final double m11 = 21.24f;
 
         final Matrix2x2 matrix = new Matrix2x2(m00, m01, m10, m11);
 
-        final float actual = matrix.determinant();
-        final float expected = m00 * m11 - m01 * m10;
+        final double actual = matrix.determinant();
+        final double expected = m00 * m11 - m01 * m10;
 
         Assert.assertEquals(expected, actual, EPSILON);
     }
 
     @Test
     public void testMultiply() {
-        final float m00 = 10.43f;
-        final float m01 = -3.58f;
-        final float m10 = 34.5f;
-        final float m11 = 201.24f;
+        final double m00 = 10.43f;
+        final double m01 = -3.58f;
+        final double m10 = 34.5f;
+        final double m11 = 201.24f;
 
-        final float n00 = 2.44f;
-        final float n10 = 1.01f;
+        final double n00 = 2.44f;
+        final double n10 = 1.01f;
 
         final Matrix2x2 M = new Matrix2x2(m00, m01, m10, m11);
         final Matrix2x1 N = new Matrix2x1(n00, n10);
@@ -44,13 +44,13 @@ public class Matrix2x2Test {
 
     @Test
     public void testInvertedSingularMatrix() {
-        final float m00 = -1.3f;
-        final float m01 = 0.0f;
-        final float m10 = 3.33f;
-        final float m11 = 0.0f;
+        final double m00 = -1.3f;
+        final double m01 = 0.0f;
+        final double m10 = 3.33f;
+        final double m11 = 0.0f;
 
         final Matrix2x2 singular = new Matrix2x2(m00, m01, m10, m11);
-        final float actual = singular.determinant();
+        final double actual = singular.determinant();
 
         Assert.assertEquals(0.0f, actual, EPSILON);
         Assert.assertTrue(singular.inverted().isEmpty());
@@ -58,10 +58,10 @@ public class Matrix2x2Test {
 
     @Test
     public void testInverted() {
-        final float m00 = 10.43f;
-        final float m01 = -3.58f;
-        final float m10 = 34.5f;
-        final float m11 = 201.24f;
+        final double m00 = 10.43f;
+        final double m01 = -3.58f;
+        final double m10 = 34.5f;
+        final double m11 = 201.24f;
 
         final Matrix2x2 matrix = new Matrix2x2(m00, m01, m10, m11);
 
@@ -70,7 +70,8 @@ public class Matrix2x2Test {
         Assert.assertTrue(inverse.isPresent());
 
         final Matrix2x2 actual = inverse.get();
-        final Matrix2x2 expected = new Matrix2x2(0.090548995807856866712f, 0.0016108398180884892806f,
+        final Matrix2x2 expected = new Matrix2x2(
+                0.090548995807856866712f, 0.0016108398180884892806f,
                 -0.01552345634750080452f, 0.0046930333247661852507f);
 
         Assert.assertEquals(expected, actual);
